Refactor honor tracking and calculation services to enhance data handling and user experience. This update includes improvements to the HonorTrackingController, HonorCriterion model, and various honor calculation services, ensuring better alignment with backend changes. Additionally, frontend assets have been optimized for improved performance across the application.

+++ app/Models/HonorCriterion.php
@@ -28,8 +28,8 @@ class HonorCriterion extends Model
     protected $casts = [
         'min_gpa' => 'float',
         'max_gpa' => 'float',
-        'min_grade' => 'integer',
-        'min_grade_all' => 'integer',
+        'min_grade' => 'float',  // Changed to float to support SHS 1.0-5.0 grading scale
+        'min_grade_all' => 'float',  // Changed to float to support SHS 1.0-5.0 grading scale
         'min_year' => 'integer',
         'max_year' => 'integer',
         'require_consistent_honor' => 'boolean',
